Started GUI/server integration
I created a GUI for the start of the game for the creation/joining of a server. Server creation guaranteed works, client joining has yet to be implemented (but the error message works, at least).  I put some guards on the input to steer people towards entering a proper IP format (i.e. only numbers and period are accepted), but I have no other guards preventing people from entering straight numbers or periods or the like...